feat: add main for console
