+++ examples/basic.rs
@@ -1,20 +1,33 @@
-mod runtime;
-
 use deno_core::anyhow::Result;
 use deno_core::error::AnyError;
+use deno_core::op;
 use deno_core::RuntimeOptions;
 use deno_core::{Extension, JsRuntime};
 use std::rc::Rc;
 
+#[op]
+pub async fn console_error(s: String) {
+    print!("\x1b[0;31m{}\x1b[0m", s);
+}
+
+#[op]
+pub async fn console_warn(s: String) {
+    print!("\x1b[93;40m{}\x1b[0m", s);
+}
+
+#[op]
+pub async fn console_ok(s: String) {
+    print!("\x1b[0;32m{}\x1b[0m", s);
+}
 // 运行js文件
 pub async fn run_js_file(filepath: &str) -> Result<(), AnyError> {
     let main_module = deno_core::resolve_path(filepath)?;
     // 扩展
     let extension = Extension::builder()
         .ops(vec![
-            runtime::console_error::decl(),
-            runtime::console_warn::decl(),
-            runtime::console_ok::decl(),
+            console_error::decl(),
+            console_warn::decl(),
+            console_ok::decl(),
         ])
         .build();
     let mut js_runtime = JsRuntime::new(RuntimeOptions {
